+++ packages/input-autocomplete/src/autocomplete-field/Component.tsx
@@ -6,6 +6,7 @@ import { InputProps } from '@alfalab/core-components-input';
 import { InputDesktop as DefaultInput } from '@alfalab/core-components-input/desktop';
 import type { FieldProps } from '@alfalab/core-components-select/shared';
 
+import { OnInputReason } from '../enums';
 import { InputAutocompleteCommonProps } from '../types';
 
 import styles from './index.module.css';
@@ -48,7 +49,8 @@ export const AutocompleteField = ({
         [onClick],
     );
 
-    const handleInput: InputProps['onChange'] = (_, payload) => onInput?.(payload.value);
+    const handleInput: InputProps['onChange'] = (_, payload) =>
+        onInput?.(payload.value, OnInputReason.Change);
 
     return (
         <Input