+++ packages/input-autocomplete/src/enums.ts
@@ -1,0 +1,4 @@
+export enum OnInputReason {
+    Close = 'close',
+    Change = 'change',
+}