+++ packages/pure-cell/src/components/graphics/component.tsx
@@ -29,6 +29,16 @@ export type Props = {
      * Отступ от графики
      */
     graphicPadding?: 'airy' | 'default' | 'compact' | 'tiny' | 'none';
+
+    /**
+     * Клик по контенту графики.
+     */
+    onClick?: () => void;
+};
+
+const GRAPHICS_COMPONENT: Record<string, keyof Pick<React.ReactHTML, 'button' | 'section'>> = {
+    button: 'button',
+    section: 'section',
 };
 
 export const Graphics: React.FC<Props> = ({
@@ -36,22 +46,43 @@ export const Graphics: React.FC<Props> = ({
     dataTestId,
     verticalAlign = 'top',
     graphicPadding,
+    onClick,
 }) => {
     const pureCellContext = useContext(PureCellContext);
 
+    const Component = onClick ? GRAPHICS_COMPONENT.button : GRAPHICS_COMPONENT.section;
+
     const defaultGraphicPadding = pureCellContext.direction === 'horizontal' ? 'airy' : 'default';
 
+    const onMouseEvents = {
+        onClick: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+            onClick?.();
+        },
+        onMouseEnter: () => {
+            pureCellContext.unsetMainHover?.();
+        },
+        onMouseLeave: () => {
+            pureCellContext.setMainHover?.();
+        },
+        onMouseDown: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+        },
+    };
+
     return (
-        <section
+        <Component
             className={cn(
                 styles.component,
                 styles[verticalAlign],
                 styles[pureCellContext.direction || 'horizontal'],
                 styles[graphicPadding || defaultGraphicPadding],
+                { [styles.button]: onClick },
             )}
             data-test-id={getDataTestId(dataTestId || pureCellContext.dataTestId, 'graphics')}
+            {...(onClick && onMouseEvents)}
         >
             {children}
-        </section>
+        </Component>
     );
 };