+++ packages/pure-cell/src/components/main/component.tsx
@@ -29,24 +29,59 @@ type Props = {
      * Используется модификатор -main
      */
     dataTestId?: string;
+
+    /**
+     * Клик по контенту.
+     */
+    onClick?: () => void;
 };
 
-export const Main: React.FC<Props> = ({ children, isReverse, className, dataTestId }) => {
-    const { direction = 'horizontal', dataTestId: contextDataTestId } = useContext(PureCellContext);
+const MAIN_COMPONENT: Record<string, keyof Pick<React.ReactHTML, 'button' | 'section'>> = {
+    button: 'button',
+    section: 'section',
+};
+
+export const Main: React.FC<Props> = ({ children, isReverse, className, dataTestId, onClick }) => {
+    const {
+        direction = 'horizontal',
+        dataTestId: contextDataTestId,
+        setMainHover,
+        unsetMainHover,
+    } = useContext(PureCellContext);
+
+    const Component = onClick ? MAIN_COMPONENT.button : MAIN_COMPONENT.section;
+
+    const onMouseEvents = {
+        onClick: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+            onClick?.();
+        },
+        onMouseEnter: () => {
+            unsetMainHover?.();
+        },
+        onMouseLeave: () => {
+            setMainHover?.();
+        },
+        onMouseDown: (e: React.MouseEvent<HTMLButtonElement>) => {
+            e.stopPropagation();
+        },
+    };
 
     return (
-        <div
+        <Component
             className={cn(
                 styles.component,
                 styles[direction],
                 {
                     [styles.reverse]: isReverse,
                 },
+                { [styles.button]: onClick },
                 className,
             )}
             data-test-id={getDataTestId(dataTestId || contextDataTestId, 'main')}
+            {...(onClick && onMouseEvents)}
         >
             {children}
-        </div>
+        </Component>
     );
 };