+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 11.4.2
+
+### Patch Changes
+
+<sup><time>27.04.2024</time></sup>
+
+### [#1174](https://github.com/core-ds/core-components/pull/1174)
+
+-   Обновление библиотеки ui-primitives и зависящих от нее файлов.
+
 ## 11.4.1
 
 ### Patch Changes